+++ src/main/java/com/crosscert/firewall/service/LoginService.java
@@ -1,0 +1,44 @@
+package com.crosscert.firewall.service;
+
+import com.crosscert.firewall.annotation.LogTrace;
+import com.crosscert.firewall.dto.MemberDTO;
+import com.crosscert.firewall.entity.Ip;
+import com.crosscert.firewall.entity.Member;
+import com.crosscert.firewall.entity.Role;
+import com.crosscert.firewall.repository.MemberRepository;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.log4j.Log4j2;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
+import org.springframework.security.core.userdetails.User;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UserDetailsService;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
+import org.springframework.security.crypto.password.PasswordEncoder;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+import java.util.ArrayList;
+import java.util.List;
+
+@Service
+@RequiredArgsConstructor
+@Transactional
+@Log4j2
+public class LoginService implements UserDetailsService {
+
+    private final MemberRepository memberRepository;
+
+    //스프링시큐리티 로그인
+    @Override
+    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
+        Member member = memberRepository.findByEmail(username).orElseThrow(() -> new UsernameNotFoundException("없는 회원 정보 입니다."));
+        List<GrantedAuthority> authorities = new ArrayList<>();
+        authorities.add(new SimpleGrantedAuthority(member.getRole().name()));
+        return new User(member.getEmail(), member.getPassword(), authorities);
+    }
+}
+
+
+
+
